module_8/app/requirements.txt
Pillow==7.1.2
numpy==1.20.2
tensorflow==2.5.0